Extract 'getBasenamesAndZipBuffers' to lib/util
