--- lib/util/zipLambdaFileToBuffer.js
@@ -1,7 +1,0 @@
-const childProcess = require("child_process");
-
-const zipLambdaFileToBuffer = (fileBasename) => {
-  return childProcess.execSync(`zip -j - lambdas/${fileBasename}.js`);
-};
-
-module.exports = zipLambdaFileToBuffer;
